fix(list-vip-requests): change from type long to int

+++ src/main/java/studyweb/cus/dto/response/admin/VipRequestListResponse.java
@@ -3,4 +3,4 @@ package studyweb.cus.dto.response.admin;
 import org.springframework.data.domain.Page;
 
 public record VipRequestListResponse(
-    Page<VipRequestResponse> requests, long totalCount, long waitingCount) {}
+    Page<VipRequestResponse> requests, int totalCount, int waitingCount) {}
